insert new product in database with SQL
