+++ website/forms.py
@@ -13,4 +13,4 @@ class ProductForm(forms.ModelForm):
 
     class Meta:
         model = Product
-        fields = ('title', 'description', 'price', 'quantity',)
+        fields = ('title', 'description', 'product_type', 'price', 'quantity',)
