+++ website/templates/product/create.html
@@ -1,19 +1,15 @@
-{% load staticfiles %}
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Bangazon</title>
-  </head>
-  <body>
-    <h1>Sell Product</h1>
+{% extends 'index.html' %}
 
+{% block content %}
+
+    <h1>Sell Product</h1>
 
     <form action="{% url 'website:sell' %}" method="post">
       {% csrf_token %}
       {{ product_form.as_p }}
       <input type="submit" value="Sell It" />
     </form>
-  </body>
-</html>
+
+{% endblock content %}
+
