+++ website/views/__init__.py
@@ -1,1 +1,2 @@
 from .index_views import *
+from .sell_views import *
